[clang] fixup generation to avoid precompilation errors, and simplify script
