+++ gen/generator.toml
@@ -1,6 +1,6 @@
 [general]
 library_name = "soapysdr"
-output_file_path = "../src/wrapper.jl"
+output_file_path = "../src/lowlevel/auto_wrap.jl"
 extract_c_comment_style = "raw"
 wrap_variadic_function = true
 print_using_CEnum = false
